Got quests working again. - The call to GetValDictionary in LocationManager.Init is called before ZoneSystem.instance.m_locationInstances has any data in it. Now when FindClosestLocation is called if m_locationInstances.Count == 0 then GetValDictionary is called again. This seems to fix the 'Try Agian,the dice is broken' error see in the LogOutput.log - Remove item from OdinData.Data.SearchTaskList when turned in. - Resolved an NRE when SearchTaskList has items in it that there is no active quest for. - Fixed Typos - Removed redundant code - I tested each quest type and they work, however the hunt quest still seems buggy and does not spawn the target if the game is closed and restarted.
