+++ .gitignore
@@ -348,3 +348,5 @@ MigrationBackup/
 
 # Ionide (cross platform F# VS Code tools) working folder
 .ionide/
+/OdinPlus-798-0-2-5-1617233860.zip
+/OdinPlus.dll
