+++ OdinPlus/5Quest/LegacyChest.cs
@@ -1,15 +1,13 @@
-using System.IO.Pipes;
-using System;
-using System.Collections.Generic;
 using UnityEngine;
+
 namespace OdinPlus
 {
-	public class LegacyChest : MonoBehaviour
+  public class LegacyChest : MonoBehaviour
 	{
 
 		private ZNetView m_nview;
-		public bool Placing = false;
-		public bool m_sphy = false;
+		public bool Placing;
+		public bool m_sphy;
 		//upd maybe make this private box??public bool isPublic = false;
 		public string m_id = "";
 		public string m_ownerName = "";
@@ -27,15 +25,13 @@ namespace OdinPlus
 				zdo.Set("QuestOwener", m_ownerName);
 				return;
 			}
-			else
+
+      m_id = zdo.GetString("QuestID", "public");
+      m_sphy = zdo.GetBool("QuestSphy", true);
+      m_ownerName = zdo.GetString("QuestOwener", "public");
+      if (!m_sphy)
 			{
-				m_id = zdo.GetString("QuestID", "public");
-				m_sphy = zdo.GetBool("QuestSphy", true);
-				m_ownerName = zdo.GetString("QuestOwener", "public");
-			}
-			if (!m_sphy)
-			{
-				DestroyImmediate(this.GetComponent<StaticPhysics>());
+				DestroyImmediate(GetComponent<StaticPhysics>());
 			}
 			/* 			if (ZNet.instance.IsServer() && ZNet.instance.IsDedicated())
 						{
@@ -69,7 +65,7 @@ namespace OdinPlus
 				if (quest != null)
 				{
 					//upd should select in base? yes!!!!!!!!!
-					QuestProcesser.Create(quest).Finish();
+					QuestProcessor.Create(quest).Finish();
 					return;
 				}
 				//upd giveup without destroy?
@@ -95,24 +91,24 @@ namespace OdinPlus
 			Collider[] array = Physics.OverlapBox(pos, Vector3.one * p_range);
 			foreach (var col in array)
 			{
-				var ctn = col.GetComponent<Container>();
-				var ctn2 = col.transform.parent.GetComponent<Container>();
-				var ctn3 = col.transform.parent.parent.GetComponent<Container>();
-				if (ctn)
+				Container ctn = col.GetComponent<Container>();
+				Container ctn2 = col.transform?.parent?.GetComponent<Container>();
+				Container ctn3 = col.transform?.parent?.parent?.GetComponent<Container>();
+				if (ctn != null)
 				{
 					ctn.gameObject.GetComponent<ZNetView>().Destroy();
 					DBG.blogWarning("Find destroyable ctn");
 					return;
 				}
-				if (ctn2)
+				if (ctn2 != null)
 				{
-					col.transform.parent.GetComponent<ZNetView>().Destroy();
+					col.transform?.parent?.GetComponent<ZNetView>().Destroy();
 					DBG.blogWarning("Find destroyable ctn parent");
 					return;
 				}
-				if (ctn3)
+				if (ctn3 != null)
 				{
-					ctn3.GetComponent<ZNetView>().Destroy();
+					ctn3?.GetComponent<ZNetView>().Destroy();
 					DBG.blogWarning("Find destroyable ctn grandparent");
 					return;
 				}
@@ -125,7 +121,7 @@ namespace OdinPlus
 		public static GameObject Place(Vector3 pos, string p_id, string p_owner, int p_key, Quaternion rot, bool sphy = true)
 		{
 			GameObject chest;
-			chest = Instantiate(ZNetScene.instance.GetPrefab("LegacyChest" + (p_key + 1).ToString()), pos, rot, OdinPlus.PrefabParent.transform);
+			chest = Instantiate(ZNetScene.instance.GetPrefab("LegacyChest" + (p_key + 1)), pos, rot, OdinPlus.PrefabParent.transform);
 
 			var lc = chest.GetComponent<LegacyChest>();
 			lc.Placing = true;
